requirements.txt
a2wsgi==1.10.8
ansio==0.0.1
browser-use==0.2.5
docker==7.1.0
duckduckgo-search==6.1.12
faiss-cpu==1.11.0
fastmcp==2.3.4
flask[async]==3.0.3
flask-basicauth==0.2.0
flaredantic==0.1.4
GitPython==3.1.43
inputimeout==1.0.4
kokoro>=0.9.2
simpleeval==1.0.3
langchain-core==0.3.49
langchain-community==0.3.19
langchain-unstructured[all-docs]==0.1.6
openai-whisper==20240930
lxml_html_clean==0.3.1
markdown==3.7
mcp==1.9.0
newspaper3k==0.2.8
paramiko==3.5.0
playwright==1.52.0
pypdf==4.3.1
python-dotenv==1.1.0
pytz==2024.2
sentence-transformers==3.0.1
tiktoken==0.8.0
unstructured[all-docs]==0.16.23
unstructured-client==0.31.0
webcolors==24.6.0
nest-asyncio==1.6.0
crontab==1.0.1
litellm==1.74
markdownify==1.1.0
pymupdf==1.25.3
pytesseract==0.3.13
pdf2image==1.17.0
crontab==1.0.1
pathspec>=0.12.1
psutil>=7.0.0
soundfile==0.13.1
pytest-asyncio==0.23.5
